Revert "move dequantize_weight() to PackableQuantLinear"
This reverts commit b5d311d6b36b98032a6bd4c56151b63e14b094ae.
